requirements/local.txt
-r base.txt

# Development tools
django-debug-toolbar==4.2.0
black==23.11.0
flake8==6.1.0
pytest==7.4.3
pytest-django==4.7.0
